+++ waspc/cli/src/Wasp/Cli/AppComponentPorts.hs
@@ -13,7 +13,7 @@ import qualified Data.List.NonEmpty as NE
 import Data.Maybe (catMaybes, isJust)
 import Network.Socket (PortNumber)
 import Wasp.Cli.Command (Command, CommandError (CommandError))
-import Wasp.Cli.Port (firstAvailableLocalPort, isLocalPortTaken, maxNumOfPortsToCheck)
+import Wasp.Cli.Port (checkIfLocalPortIsTaken, findFirstFreeLocalPort, maxNumOfPortsToCheck)
 import Wasp.Util (whenM)
 
 defaultDevClientPort :: PortNumber
@@ -49,13 +49,13 @@ findAppComponentPorts (requestedClientPort, requestedServerPort) = do
   return (resolvedClientPort, resolvedServerPort)
   where
     assertPort port = do
-      whenM (liftIO $ isLocalPortTaken port) $ do
+      whenM (liftIO $ checkIfLocalPortIsTaken port) $ do
         throwResolvingError $ "Port " ++ show port ++ " is already in use."
       return port
 
     findPort startPort portsToSkip = do
       let candidatePorts = take maxNumOfPortsToCheck [startPort ..] \\ portsToSkip
-      availablePort <- liftIO (firstAvailableLocalPort candidatePorts)
+      availablePort <- liftIO (findFirstFreeLocalPort candidatePorts)
       case availablePort of
         Nothing -> throwResolvingError $ noFreePortError candidatePorts
         Just port -> return port