+++ waspc/src/Wasp/Db/Postgres.hs
@@ -5,15 +5,16 @@ module Wasp.Db.Postgres
   )
 where
 
+import Network.Socket (PortNumber)
 import Text.Printf (printf)
 import Wasp.Util.Docker
   ( DockerImageName,
     DockerVolumeMountPath,
   )
 
-makeConnectionUrl :: String -> String -> Int -> String -> String
+makeConnectionUrl :: String -> String -> PortNumber -> String -> String
 makeConnectionUrl user pass port dbName =
-  printf "postgresql://%s:%s@localhost:%d/%s" user pass port dbName
+  printf "postgresql://%s:%s@localhost:%s/%s" user pass (show port) dbName
 
 -- As specified by PostgreSQL documentation.
 postgresMaxDbNameLength :: Int